Only backup DB tablespace if leader.

+++ backup_db.sh
@@ -9,7 +9,8 @@ AKID="{\"s\": {\"opitem\": \"AWS\", \"opfield\": \"${AWS_DEFAULT_REGION}.akid\"}
 export AWS_ACCESS_KEY_ID="$(/opt/app/bin/python /opt/app/pylib/cred_tool <<< "${AKID}")"
 SAK="{\"s\": {\"opitem\": \"AWS\", \"opfield\": \"${AWS_DEFAULT_REGION}.sak\"}}"
 export AWS_SECRET_ACCESS_KEY="$(/opt/app/bin/python /opt/app/pylib/cred_tool <<< "${SAK}")"
-if [ -f "${TABLESPACE_PATH}" ]; then
+# backup only if the tablespace exists and this device is the leader
+if [ -f "${TABLESPACE_PATH}" ] && [ -f "/data/is_leader" ]; then
   aws s3 cp "${TABLESPACE_PATH}" "s3://tailucas-automation/${APP_NAME}.db" --only-show-errors
 else
   aws s3 cp "s3://tailucas-automation/${APP_NAME}.db" "${TABLESPACE_PATH}" --only-show-errors
